+++ lib/palettes.cpp
@@ -48,9 +48,9 @@ int PaletteState::ChoosePalette()
    {
       pal = 1;
    }
-   else if (pal > (int)Palettes->NumPalettes())
+   else if (pal > Palettes->NumPalettes())
    {
-      pal = (int)Palettes->NumPalettes();
+      pal = Palettes->NumPalettes();
    }
 
    return pal-1;
@@ -202,8 +202,7 @@ void PaletteState::ToTextureSmooth(double * palette, size_t plt_size,
 PaletteState::PaletteState()
    : first_init(false)
    ,Palettes(&BasePalettes)
-   , palette_tex(BasePalettes.NumPalettes()
-                )
+   , palette_tex(BasePalettes.NumPalettes())
 {}
 
 static std::mutex init_mtx;
@@ -229,7 +228,7 @@ void PaletteState::Init()
       glGenTextures(Palettes->NumPalettes() * 2, &(paletteTexIds[0][0]));
       glGenTextures(1, &alphaTexId);
 
-      for (size_t ipal = 0; ipal < Palettes->NumPalettes(); ipal++)
+      for (int ipal = 0; ipal < Palettes->NumPalettes(); ipal++)
       {
          palette_tex[ipal][0] = paletteTexIds[ipal][0];
          palette_tex[ipal][1] = paletteTexIds[ipal][1];
@@ -266,7 +265,7 @@ void PaletteState::Init()
       first_init = true;
    }
 
-   for (int i = 0; i < (int)Palettes->NumPalettes(); i++)
+   for (int i = 0; i < Palettes->NumPalettes(); i++)
    {
       ToTextureDiscrete(GetData(i),
                         GetSize(i),
